requirements.txt
aiogram==3.12.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
python-dotenv==1.0.1
aiohttp==3.10.11
certifi>=2024.0.0  # Для SSL сертификатов в aiohttp
# Pin ozonapi-async to 0.1.0 because newer releases require aiohttp>=3.13, which conflicts with aiogram<3.11; 0.1.0 works with aiohttp 3.10.x.
ozonapi-async==0.1.0
openai>=1.50.0
SQLAlchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0
pydantic-settings==2.12.0
python-barcode==0.15.1
Pillow==11.0.0
reportlab==4.2.2
psutil>=5.9.0
APScheduler>=3.10.4,<3.12.0  # Для JobQueue (как в python-telegram-bot)
cachetools>=5.3.3,<6.3.0  # Для CallbackDataCache